warning fixes (few of them)

+++ scheduler/nufa/src/nufa.c
@@ -43,21 +43,21 @@ nufa_init (struct xlator *xl)
   trav_xl = xl->first_child;
   index = 0;
   
-  *((int *)xl->private) = (int)nufa_buf; // put it at the proper place
+  *((long *)xl->private) = (long)nufa_buf; // put it at the proper place
   return 0;
 }
 
 static void
 nufa_fini (struct xlator *xl)
 {
-  struct nufa_struct *nufa_buf = (struct nufa_struct *)*((int *)xl->private);
+  struct nufa_struct *nufa_buf = (struct nufa_struct *)*((long *)xl->private);
   free (nufa_buf);
 }
 
 static struct xlator *
 nufa_schedule (struct xlator *xl, int size)
 {
-  struct nufa_struct *nufa_buf = (struct nufa_struct *)*((int *)xl->private);
+  struct nufa_struct *nufa_buf = (struct nufa_struct *)*((long *)xl->private);
   return nufa_buf->sched_xl;
 }
